mini-projet/auth-service/requirements.txt
flask
flask_cors
pyjwt
